pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wvcr"
version = "0.1.0"
authors = [
    {name = "Your Name"}
]
description = "Voice recording and transcription tool"
requires-python = ">=3.9"
dependencies = [
    "pyaudio",
    "loguru",
    "openai>=1.75.0",
    "wave",
    "pyperclip",
    "pynput",
    "python-dotenv",
    "plyer",
    "requests",
    "yt-dlp",
  "ffmpeg-python",
  "hydra-core>=1.3.2"
]

[project.optional-dependencies]
wayland = ["evdev"]

[project.scripts]
# Hydra @hydra.main based CLI (modularized)
wvcr = "wvcr.cli.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["wvcr"]

[tool.pytest.ini_options]
pythonpath = [
  "src",
  "."
]
